% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{bulk.calc_SOM}
\alias{bulk.calc_SOM}
\title{Bulk calculate graph properties for CONN.}
\usage{
bulk.calc_SOM(
  source.df,
  supervised = FALSE,
  save_to.excel = FALSE,
  method = "allPEs"
)
}
\arguments{
\item{source.df}{The data frame containing directories and related info about the
CONN matrix from multiple stages. The detailed information of the dataframe
is referring to \code{\link[=bulk.parm_SOM]{bulk.parm_SOM()}}.}

\item{supervised}{The choice of with or w/o supervised measurements in the calculation.}

\item{save_to.excel}{The choice for save all properties to an .xlsx file in the current working directory.}

\item{method}{The choice for taking \code{allPEs}: all PEs or \code{activePEs}: active PEs into account.}
}
\value{
A list with graph measurement results.

\code{graph.property} A total of 31 Unsupervised properties and 1 supervised property of the CONN graph

\code{graph.degree} Degree vector of the CONN graph

\code{graph.eigenvalue} Eigenvalue vector of the CONN graph
}
\description{
This function create graph objects and futher calculate supervised and unsupervised
graph properties that are adaptive to CONN graph for multiple stages of CONN matrix.
}
\seealso{
\link[=calc.graph_property]{load.graph_SOM()} for the single stage version of calculation
}
