% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc.graph_property}
\alias{calc.graph_property}
\title{Calculate graph properties}
\usage{
calc.graph_property(
  list,
  supervised = FALSE,
  save_to.excel = FALSE,
  method = "allPEs"
)
}
\arguments{
\item{save_to.excel}{The choice for save all to an .xlsx file in the subfolder '/res'}

\item{method}{The choice for taking "allPEs": all PEs \ "activePEs": active PEs into account}

\item{conn.graph}{A CONN graph}

\item{param.table}{The key-value pair parameter table}
}
\value{
A list with graph measurement results.
}
\description{
This function calculate several graph properties for CONN graph.
}
