% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.graph_SOM}
\alias{load.graph_SOM}
\title{Load CONN matrix as graph object}
\usage{
load.graph_SOM(source.df, method = "weighted")
}
\arguments{
\item{source.df}{The product data frame from func:load.parm_SOM.}

\item{method}{Set the graph type: "weighted": weighted graph/ "binary": binary graph}
}
\value{
A list consists of both the graph object and the related parameter data frame
}
\description{
This function built a graph for the CONN matrix based on the related SOM products
accessing from directory link.
}
