% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load.parm_SOM}
\alias{load.parm_SOM}
\title{Load parameter table.}
\usage{
load.parm_SOM(infile)
}
\arguments{
\item{infile}{Path to the input .cadj file}
}
\value{
A data frame with only one row showing the current learning steps and
directories of related SOM files.
}
\description{
This function search realated possible wwightcube .wgtcub/.wgt.xv file and
the class label .nunr file for the input list of CONN .cadj file for each
learning steps. And integrated as an output dataframe.
}
