% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{bulk.calc_SOM}
\alias{bulk.calc_SOM}
\title{Bulk load parameter table.}
\usage{
bulk.calc_SOM(
  source.df,
  supervised = FALSE,
  save_to.excel = FALSE,
  method = "allPEs"
)
}
\arguments{
\item{infile}{Path to the input .parm/.txt file with the directories of .cadj file}
}
\value{
A data frame with each row showing the current learning steps and
directories of related SOM files.
}
\description{
This function search realated possible wwightcube .wgtcub/.wgt.xv file and
the class label .nunr file for the input list of CONN .cadj file for each
learning steps. And integrated as an output dataframe.
}
